separate out stable map and set to prevent access to internals
