+++ cranelift/isle/isle/src/sema.rs
@@ -17,7 +17,7 @@ use crate::ast;
 use crate::error::*;
 use crate::lexer::Pos;
 use crate::log;
-use crate::{StableMap, StableSet};
+use crate::stablemapset::{StableMap, StableSet};
 use std::collections::hash_map::Entry;
 use std::collections::BTreeMap;
 use std::collections::BTreeSet;
